Added code to stop the creation of core files. Changed all references to log() to log_message().

+++ src/tinyproxy.h
@@ -1,4 +1,4 @@
-/* $Id: tinyproxy.h,v 1.7 2001-05-23 18:00:54 rjkaes Exp $
+/* $Id: tinyproxy.h,v 1.8 2001-05-27 02:36:22 rjkaes Exp $
  *
  * See 'tinyproxy.c' for a detailed description.
  *
@@ -62,7 +62,7 @@
 #endif
 
 /* Global variables for the main controls of the program */
-#define MAXBUFFSIZE	(1024 * 48)	/* Max size of buffer */
+#define MAXBUFFSIZE	(size_t)(1024 * 48)	/* Max size of buffer */
 #define MAXLISTEN	1024		/* Max number of connections */
 #define MAX_IDLE_TIME 	(60 * 10)	/* 10 minutes of no activity */
